% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_tpl_to_date}
\alias{apply_tpl_to_date}
\title{wrapper um glue:glue and format}
\usage{
apply_tpl_to_date(date, tpl, ...)
}
\arguments{
\item{date}{von der generierten Sequenz (POSIXct oder Date)}

\item{tpl}{Template String aus der config}

\item{...}{Möglichkeit weitere Parameter für \code{\link[glue:glue]{glue::glue()}} einzufügen}
}
\value{
formatierter String
}
\description{
Dadurch das es ein Wrapper um \code{\link[glue:glue]{glue::glue()}} und \code{\link[base:format]{base::format()}} ist, kann
man zum einem das Datum einfach mit den üblichen Formatierungskürzel wie \code{"\%Y"}
formatieren oder Ausdrücke einfügen.
}
\examples{
to <- lubridate::floor_date(lubridate::now("GMT"), "day")
from <- to - lubridate::period("6 days")
dates <- seq(from, to, by = "1 day")

tpl <- "[..]/mythenquai?startDate={format(date - 1, '\%Y-\%m-\%d')}&endDate=\%Y-\%m-\%d"

apply_tpl_to_date(dates, tpl)

}
