% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_date_sequence}
\alias{create_date_sequence}
\title{Zeitpunkte für die Dateinamen generieren}
\usage{
create_date_sequence(config)
}
\arguments{
\item{config}{Konfiguration für gesamten import}
}
\value{
Vector von Date oder POSIXct
}
\description{
Generiert eine Sequence von Zeitpunkt aus der config
}
\examples{
config_file <- system.file("extdata", "config_mythenquai.yml", package = "rOstluft.import")
config <- yaml::read_yaml(config_file)

create_date_sequence(config)
}
