% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_plan_df}
\alias{create_plan_df}
\title{Erzeuge Dataframe aller Dateien}
\usage{
create_plan_df(dates, config)
}
\arguments{
\item{dates}{Sequence von Date/POSIXct}

\item{config}{Konfiguration f端r gesamten import}
}
\value{
tibble mit allen notwendigen Angaben f端r den Import
}
\description{
Erzeugt einen Dataframe mit einer Zeile f端r jedes Datum/Zeitpunkt. Folgende Angaben
ist in jeder Zeile:
}
\details{
\itemize{
\item input_fn: Input Dateiname/URL zum kopieren/herunterladen
\item cache_fn: Cache Dateiname der Rohdaten
\item imported_fn: Import Dateiname f端r die normalisierten und importierten Daten
\item cache_fn_exists: exisitiert die Cache Datei bereits?
\item imported_fn_exists: exisitiert die Import Datei bereits?
}
}
\examples{
config_file <- system.file("extdata", "config_mythenquai.yml", package = "rOstluft.import")
config <- yaml::read_yaml(config_file)

dates <- create_date_sequence(config)
plan <- create_plan_df(dates, config)

tibble::glimpse(plan)
}
