% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_store}
\alias{get_store}
\title{Erzeugt lokalen rolf rds store}
\usage{
get_store(config)
}
\arguments{
\item{config}{Konfiguration für gesamten import}
}
\value{
schreibbarer lokaler rOstluft
}
\description{
Erzeugt lokalen rolf rds store
}
\examples{
config_file <- system.file("extdata", "config_mythenquai.yml", package = "rOstluft.import")
config <- yaml::read_yaml(config_file)

get_store(config)
}
