% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Importiert Daten}
\usage{
import(config)
}
\arguments{
\item{config}{Liste mit sämtlichen Konfigurationsoptionen}
}
\value{
invisible(NULL)
}
\description{
Diese Funktionen importiert Daten gemäss einer Konfiguration in einen lokalen
rolf rds store. Damit der Import durchgeführt werden kann, müssen folgende
Anforderungen erfüllt sein:
}
\details{
\itemize{
\item Eine Lesefunktion für die Daten muss entweder als Funktion in einem Package
oder im globalen Enviroment existieren
\item Der Store enthält die notwendigen Metadaten
}

Aktuell werden als Quelle curl Downloads (ftp, http) und das Dateisystem unterstütz.
}
\section{config}{

Die Konfiguration wird am einfachsten in einer externen yaml Datei gespeichert.
Eine Beschreibung der verschiedenen Optionen ist in der Beispieldatei
\code{config_mythenquai.yml} enthalten.
}

\section{logging}{

Das Packaging nutzt einen \link[lgr:lgr-package]{lgr::lgr} Package Level Logger mit dem Namen
"rOstluft.import". Mehr Infos in der \href{https://s-fleck.github.io/lgr/articles/lgr.html}{lgr Vignette}
}

\examples{
lg <- lgr::get_logger("rOstluft.import")
lg$set_threshold("trace")

config_file <- system.file("extdata", "config_mythenquai.yml", package = "rOstluft.import")
config <- yaml::read_yaml(config_file)
import(config)
}
